feat: Begins to move `Newsletter` to fs/components
