+++ packages/ui/src/components/organisms/Newsletter/styles.scss
@@ -4,7 +4,7 @@
   // --------------------------------------------------------
 
   // Default properties
-  --fs-newsletter-padding-mobile                    : var(--fs-spacing-4) var(--fs-spacing-3) var(--fs-spacing-5);
+  --fs-newsletter-padding-mobile                    : var(--fs-spacing-5);
   --fs-newsletter-padding-desktop                   : var(--fs-spacing-9) 10%;
   --fs-newsletter-border-radius                     : 0;
 
@@ -38,39 +38,25 @@
   border-radius: var(--fs-newsletter-border-radius);
 
   [data-fs-newsletter-form] {
+    padding: var(--fs-newsletter-padding-mobile);
+    display: flex;
+    flex-direction: column;
+    row-gap: var(--fs-spacing-4);
 
     @include media(">=notebook") {
-      display: grid;
-      grid-template-columns: repeat(2, 1fr);
-      column-gap: var(--fs-spacing-8);
-      padding: var(--fs-newsletter-padding-mobile);
-
-      @include media(">=notebook") {
-        padding: var(--fs-newsletter-padding-desktop);
-      }
-
-      [data-fs-newsletter-header] {
-        align-content: flex-start;
-      }
+      padding: var(--fs-newsletter-padding-desktop);
     }
 
-    [data-fs-newsletter-controls] {
-      @include media(">=notebook") {
-        grid-template-columns: repeat(3, 1fr);
-        margin-top: 0;
+  }
 
-        [data-fs-input-field] {
-          grid-column: 1 / span 3;
-        }
+  [data-fs-newsletter-controls] {
+    display: flex;
+    flex-direction: column;
+    row-gap: var(--fs-spacing-3);
 
-        [data-fs-button-inverse] {
-          grid-column: 3;
-        }
-
-        [data-fs-newsletter-addendum] {
-          grid-column: 1 / span 3;
-        }
-      }
+    [data-fs-button] {
+      width: 100%;
+      margin-top: var(--fs-spacing-3);
     }
   }
 
@@ -82,31 +68,29 @@
 
     @include media(">=notebook") {
       text-align: left;
+      align-content: flex-start;
     }
   }
 
-  [data-fs-newsletter-header] h3 {
-    display: grid;
-    grid-template-columns: var(--fs-spacing-5) auto;
+  [data-fs-newsletter-header-title] {
+    display: flex;
+    flex-wrap: wrap;
+    row-gap: var(--fs-spacing-1);
     column-gap: var(--fs-spacing-1);
     align-items: center;
     justify-content: center;
     font-size: var(--fs-newsletter-title-size);
     font-weight: var(--fs-newsletter-title-weight);
+    line-height: 1.2;
+
+    [data-fs-icon] {
+      flex-shrink: 0;
+      width: var(--fs-spacing-5);
+      height: var(--fs-spacing-5);
+    }
 
     @include media(">=notebook") {
       justify-content: left;
-    }
-  }
-
-  [data-fs-newsletter-controls] {
-    display: grid;
-    row-gap: var(--fs-spacing-3);
-    margin-top: var(--fs-spacing-4);
-
-    [data-fs-button-inverse] {
-      width: 100%;
-      margin-top: var(--fs-spacing-3);
     }
   }
 
@@ -122,25 +106,31 @@
   // Variants Styles
   // --------------------------------------------------------
 
+  &:not([data-fs-newsletter="card"]) {
+    @include media(">=notebook") {
+      [data-fs-newsletter-form] {
+        display: grid;
+        grid-template-columns: repeat(2, 1fr);
+        column-gap: var(--fs-spacing-8);
+      }
+      [data-fs-button] {
+        width: auto;
+        margin-left: auto;
+        min-width: 40%;
+      }
+    }
+  }
+
   &[data-fs-newsletter="card"] {
     text-align: center;
     border-radius: var(--fs-newsletter-card-border-radius);
 
-    [data-fs-newsletter-header] {
+    [data-fs-newsletter-header],
+    [data-fs-newsletter-addendum] {
       text-align: center;
     }
 
-    @include media(">=notebook") {
-      [data-fs-newsletter-form] {
-        display: block;
-      }
-
-      [data-fs-newsletter-controls] {
-        margin-top: var(--fs-spacing-4);
-      }
-    }
-
-    [data-fs-newsletter-header] h3 {
+    [data-fs-newsletter-header-title] {
       justify-content: center;
     }
   }
